move resource manager tests into correct dir
