+++ packages/dashboard/src/managers/tests/resource-manager.test.tsx
@@ -1,5 +1,5 @@
-import ResourceManager from '../managers/resource-manager';
-import fakeResources from '../managers/__mocks__/resources';
+import ResourceManager from '../resource-manager';
+import fakeResources from '../__mocks__/resources';
 
 describe('Correct instantiation', () => {
   test('Instantiate robots and dispensers', () => {
